requirements.txt
pyrogram
tgcrypto
python-dotenv
requests
motor
aiohttp
